+++ Frontend/Validators/FeaturesInitiatedValidator.cs
@@ -1,6 +1,7 @@
 ﻿using Data.Models.Projects;
 using FluentValidation;
 using Frontend.Models;
+using Frontend.Models.Features;
 
 namespace Frontend.Validators
 {
@@ -10,15 +11,6 @@ namespace Frontend.Validators
         {
             var whoInitiatedError = "Select who initiated the project";
             RuleFor(x => x.WhoInitiated).NotEqual(TransferFeatures.ProjectInitiators.Empty).WithMessage(whoInitiatedError);
-        }
-    }
-
-    public class FeaturesReasonValidator : AbstractValidator<FeaturesReasonViewModel>
-    {
-        public FeaturesReasonValidator()
-        {
-            var subjectToInterventionError = "Select whether or not the transfer is subject to intervention";
-            RuleFor(x => x.IsSubjectToIntervention).NotNull().WithMessage(subjectToInterventionError);
         }
     }
 }
