+++ Frontend/Views/Features/Index.cshtml
@@ -1,7 +1,7 @@
 @using Data.Models.Projects
 @using Frontend.Helpers
 @using Helpers
-@model FeaturesViewModel
+@model Frontend.Models.Features.FeaturesViewModel
 
 @{
     ViewBag.Title = "Features of the transfer";
