+++ Frontend/Views/Features/Reason.cshtml
@@ -1,5 +1,5 @@
 @using Microsoft.AspNetCore.Mvc.ModelBinding
-@model Frontend.Models.FeaturesReasonViewModel
+@model Frontend.Models.Features.FeaturesReasonViewModel
 
 @{
     ViewBag.Title = !ViewData.ModelState.IsValid ? "Error: Reason for transfer" : "Reason for transfer";
