+++ Frontend/Views/Features/Type.cshtml
@@ -1,5 +1,5 @@
 @using Data.Models.Projects
-@model FeaturesViewModel
+@model Frontend.Models.Features.FeaturesViewModel
 
 @{
     ViewBag.Title = Model.FormErrors.ProcessPageTitleIfThereIsAnError("What type of transfer is it?");
